+++ Sources/Scout/Extensions/AEXML/AEXMLElement+Children.swift
@@ -13,6 +13,17 @@ extension AEXMLElement {
         AEXMLElement(name: defaultArrayElementName, value: value)
     }
 
+    /// Returns a new element with new children, keeping the name, value and attributes
+    func copy() -> AEXMLElement {
+        let newElement = AEXMLElement(name: name, value: value, attributes: attributes)
+        if children.isEmpty {
+            return newElement
+        }
+
+        children.forEach { newElement.addChild($0.copyFlat()) }
+        return newElement
+    }
+
     /// Copy of the element, without the children
     func copyFlat() -> AEXMLElement { AEXMLElement(name: name, value: value, attributes: attributes) }
