+++ Sources/Scout/Models/PathExplorer/PathExplorers.swift
@@ -5,10 +5,11 @@
 
 import Foundation
 
+/// Namespace to find all PathExplorers in a single place
 public enum PathExplorers {
 
-    public typealias Json = CodableFormatPathExplorer<CodableFormats.JsonDefault>
-    public typealias Plist = CodableFormatPathExplorer<CodableFormats.PlistDefault>
-    public typealias Yaml = CodableFormatPathExplorer<CodableFormats.YamlDefault>
+    public typealias Json = CodablePathExplorer<CodableFormats.JsonDefault>
+    public typealias Plist = CodablePathExplorer<CodableFormats.PlistDefault>
+    public typealias Yaml = CodablePathExplorer<CodableFormats.YamlDefault>
     public typealias Xml = ExplorerXML
 }
